+++ exporters/trace/jaeger/jaeger.go
@@ -328,21 +328,25 @@ func spanSnapshotToThrift(ss *export.SpanSnapshot) *gen.Span {
 
 	var refs []*gen.SpanRef
 	for _, link := range ss.Links {
+		tid := link.TraceID()
+		sid := link.SpanID()
 		refs = append(refs, &gen.SpanRef{
-			TraceIdHigh: int64(binary.BigEndian.Uint64(link.TraceID[0:8])),
-			TraceIdLow:  int64(binary.BigEndian.Uint64(link.TraceID[8:16])),
-			SpanId:      int64(binary.BigEndian.Uint64(link.SpanID[:])),
+			TraceIdHigh: int64(binary.BigEndian.Uint64(tid[0:8])),
+			TraceIdLow:  int64(binary.BigEndian.Uint64(tid[8:16])),
+			SpanId:      int64(binary.BigEndian.Uint64(sid[:])),
 			RefType:     gen.SpanRefType_FOLLOWS_FROM,
 		})
 	}
 
+	tid := ss.SpanContext.TraceID()
+	sid := ss.SpanContext.SpanID()
 	return &gen.Span{
-		TraceIdHigh:   int64(binary.BigEndian.Uint64(ss.SpanContext.TraceID[0:8])),
-		TraceIdLow:    int64(binary.BigEndian.Uint64(ss.SpanContext.TraceID[8:16])),
-		SpanId:        int64(binary.BigEndian.Uint64(ss.SpanContext.SpanID[:])),
+		TraceIdHigh:   int64(binary.BigEndian.Uint64(tid[0:8])),
+		TraceIdLow:    int64(binary.BigEndian.Uint64(tid[8:16])),
+		SpanId:        int64(binary.BigEndian.Uint64(sid[:])),
 		ParentSpanId:  int64(binary.BigEndian.Uint64(ss.ParentSpanID[:])),
 		OperationName: ss.Name, // TODO: if span kind is added then add prefix "Sent"/"Recv"
-		Flags:         int32(ss.SpanContext.TraceFlags),
+		Flags:         int32(ss.SpanContext.TraceFlags()),
 		StartTime:     ss.StartTime.UnixNano() / 1000,
 		Duration:      ss.EndTime.Sub(ss.StartTime).Nanoseconds() / 1000,
 		Tags:          tags,